+++ src/ca/goldorion/mcrpcreator/io/export/OutputBlock.java
@@ -1,36 +1,17 @@
-package ca.goldorion.mcrpcreator.io;
+package ca.goldorion.mcrpcreator.io.export;
 
-import ca.goldorion.mcrpcreator.MainApp;
-import ca.goldorion.mcrpcreator.io.export.OutputProcedureBlock;
-import ca.goldorion.mcrpcreator.io.export.ProceduralProcedureBlock;
 import ca.goldorion.mcrpcreator.io.jsons.*;
 import ca.goldorion.mcrpcreator.models.BlockModel;
 import ca.goldorion.mcrpcreator.utils.FileUtils;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import javafx.stage.FileChooser;
 
 import java.io.File;
 import java.util.ArrayList;
 
-public class Export {
+public class OutputBlock {
 
-    public static void export(MainApp mainApp, BlockModel selectedBlock){
-        if(selectedBlock.getBlockType().equals("Output Block")){
-            OutputProcedureBlock.outputProcedureBlock(mainApp, selectedBlock);
-        } else if(selectedBlock.getBlockType().equals("Procedural Block")){
-            ProceduralProcedureBlock.proceduralProcedureBlock(mainApp, selectedBlock);
-        }
-
-
-    }
-
-    private static void proceduralProcedureBlock(MainApp mainApp, BlockModel selectedBlock) {
-        FileChooser fileChooser = new FileChooser();
-        FileChooser.ExtensionFilter extensionFilter = new FileChooser.ExtensionFilter(
-                "JSON files", "*.json");
-        fileChooser.getExtensionFilters().add(extensionFilter);
-        File file = fileChooser.showSaveDialog(mainApp.getPrimaryStage());
+    public static void outputProcedureBlock(BlockModel selectedBlock, File file){
         if(file != null) {
             if(!file.getPath().endsWith(".json")) {
                 file = new File(file.getPath() + ".json");
@@ -150,4 +131,5 @@ public class Export {
 
         }
     }
+
 }