+++ src/ca/goldorion/mcrpcreator/io/jsons/Infos.java
@@ -1,0 +1,15 @@
+package ca.goldorion.mcrpcreator.io.jsons;
+
+public class Infos {
+    private String name;
+    private String description;
+    private String version;
+    private String author;
+
+    public Infos(String name, String description, String version, String author) {
+        this.name = name;
+        this.description = description;
+        this.version = version;
+        this.author = author;
+    }
+}