+++ src/ca/goldorion/mcrpcreator/ui/BlockOverviewController.java
@@ -1,12 +1,15 @@
 package ca.goldorion.mcrpcreator.ui;
 
 import ca.goldorion.mcrpcreator.MainApp;
-import ca.goldorion.mcrpcreator.io.Export;
+import ca.goldorion.mcrpcreator.io.export.Export;
+import ca.goldorion.mcrpcreator.io.export.PluginJsonExport;
 import ca.goldorion.mcrpcreator.models.BlockModel;
+import ca.goldorion.mcrpcreator.models.PluginJsonModel;
 import ca.goldorion.mcrpcreator.utils.AlertUtils;
 import javafx.fxml.FXML;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
+import javafx.scene.control.TextField;
 
 public class BlockOverviewController {
 
@@ -14,6 +17,19 @@ public class BlockOverviewController {
     private TableView<BlockModel> blockTable;
     @FXML
     private TableColumn<BlockModel, String> fileNameColumn;
+
+    @FXML
+    private TextField pluginId;
+    @FXML
+    private TextField pluginMin;
+    @FXML
+    private TextField pluginName;
+    @FXML
+    private TextField pluginDesc;
+    @FXML
+    private TextField pluginVersion;
+    @FXML
+    private TextField pluginAuthor;
 
     private MainApp mainApp;
 
@@ -30,6 +46,30 @@ public class BlockOverviewController {
         this.mainApp = mainApp;
 
         blockTable.setItems(mainApp.getBlockData());
+    }
+
+    @FXML
+    private void handleSavePluginInfoButton(){
+        PluginJsonModel pluginJsonModel = new PluginJsonModel();
+        if(isInputValid()) {
+            pluginJsonModel.setId(pluginId.getText());
+            if(!pluginMin.getText().isEmpty()) {
+                pluginJsonModel.setMinversion(Long.parseLong(pluginMin.getText()));
+            }
+
+            pluginJsonModel.setName(pluginName.getText());
+            if(!pluginDesc.getText().isEmpty()){
+                pluginJsonModel.setDescription(pluginDesc.getText());
+            }
+            if(!pluginVersion.getText().isEmpty()){
+                pluginJsonModel.setVersion(pluginVersion.getText());
+            }
+            if(!pluginAuthor.getText().isEmpty()){
+                pluginJsonModel.setAuthor(pluginAuthor.getText());
+            }
+
+            PluginJsonExport.pluginJson(pluginJsonModel);
+        }
     }
 
     @FXML
@@ -69,5 +109,22 @@ public class BlockOverviewController {
     private void handleExportBlock(){
         BlockModel selectedBlock = blockTable.getSelectionModel().getSelectedItem();
             Export.export(mainApp, selectedBlock);
+    }
+
+    private boolean isInputValid() {
+        String message = "";
+        if(pluginId.getText() == null || pluginId.getText().length() == 0){
+            message += "The ID is not valid!\n";
+        }
+        if(pluginName.getText() == null || pluginName.getText().length() == 0){
+            message += "The name is not valid!\n";
+        }
+
+        if(message.length() == 0){
+            return true;
+        } else{
+            AlertUtils.error("Please correct invalid field(s)", message);
+            return false;
+        }
     }
 }