+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.4.0
 * **Added the possibility to create Procedural procedure blocks**
 * Added a setting to change the inputsInline
+* A JSON file is now generated when we save the block.
+* Added the possibility to create the plugin.json
 
 ## 0.3.0
 * **Added the possibility to add arguments (input value, field input and field checkbox)**